Update Participant definition in definitions.md
Removing "and has paid for a course; he/she must also have attended the course" from "*[Participants]:  a person who is registered and has paid for a course; he/she must also have attended the course" to simplify things. A course can be free, and thus the participants does not pay. We talk as well of a participant before a course takes place (thus the person didn't attend the course yet).